services/scrapy_spiders/car_scrapers/requirements.txt
# services/scrapy_spiders/car_scrapers/requirements.txt
scrapy>=2.11.0
kafka-python-ng[ujson]>=2.1.2
requests>=2.31.0
